requirements.txt
PyGithub
openai
python-dotenv
PyYAML
google-generativeai